fix: only display marker when coords exist
