+++ components/map/map-results.tsx
@@ -7,6 +7,7 @@ import { formatIntFr, formatSiret } from '#utils/helpers';
 import maplibregl, { Map } from 'maplibre-gl';
 import 'maplibre-gl/dist/maplibre-gl.css';
 import { useEffect, useRef } from 'react';
+import { checkLatLng } from './check-lat-lng';
 import './map.css';
 
 export default function MapWithResults({
@@ -35,42 +36,52 @@ export default function MapWithResults({
     });
 
     results.results.forEach((result) => {
-      var popup = new maplibregl.Popup({ offset: 25 }).setHTML(
-        `<div><strong><a href="/entreprise/${result.chemin}">${formatIntFr(
-          result.siren
-        )}</a></strong><br/>${result.nomComplet}<br/>📍${
-          result.siege.adresse
-        }</div>`
+      const coordsSiege = checkLatLng(
+        result.siege.latitude,
+        result.siege.longitude
       );
 
-      new maplibregl.Marker({ color: constants.colors.frBlue })
-        //@ts-ignore
-        .setLngLat([result.siege.longitude, result.siege.latitude])
-        .setPopup(popup)
-        //@ts-ignore
-        .addTo(map.current);
+      if (coordsSiege) {
+        const popup = new maplibregl.Popup({ offset: 25 }).setHTML(
+          `<div><strong><a href="/entreprise/${result.chemin}">${formatIntFr(
+            result.siren
+          )}</a></strong><br/>${result.nomComplet}<br/>📍${
+            result.siege.adresse
+          }</div>`
+        );
+
+        new maplibregl.Marker({ color: constants.colors.frBlue })
+          //@ts-ignore
+          .setLngLat([result.siege.longitude, result.siege.latitude])
+          .setPopup(popup)
+          //@ts-ignore
+          .addTo(map.current);
+      }
 
       result.matchingEtablissements.forEach((match) => {
         if (match.estSiege) {
           return null;
         }
 
-        var popup = new maplibregl.Popup({ offset: 25 }).setHTML(
-          `<div><strong><a href="/etablissement/${match.siret}">${formatSiret(
-            match.siret
-          )}</a></strong><br/>Etablissement secondaire de <a href="/entreprise/${
-            match.siren
-          }">${result.nomComplet}</a><br/>📍${match.adresse}</div>`
-        );
+        const coordsEtab = checkLatLng(match.latitude, match.longitude);
+        if (coordsEtab) {
+          var popup = new maplibregl.Popup({ offset: 25 }).setHTML(
+            `<div><strong><a href="/etablissement/${match.siret}">${formatSiret(
+              match.siret
+            )}</a></strong><br/>Etablissement secondaire de <a href="/entreprise/${
+              match.siren
+            }">${result.nomComplet}</a><br/>📍${match.adresse}</div>`
+          );
 
-        new maplibregl.Marker({
-          color: shouldColorZipCode ? 'yellow' : constants.colors.pastelBlue,
-        })
-          //@ts-ignore
-          .setLngLat([match.longitude, match.latitude])
-          .setPopup(popup)
-          //@ts-ignore
-          .addTo(map.current);
+          new maplibregl.Marker({
+            color: shouldColorZipCode ? 'yellow' : constants.colors.pastelBlue,
+          })
+            //@ts-ignore
+            .setLngLat([match.longitude, match.latitude])
+            .setPopup(popup)
+            //@ts-ignore
+            .addTo(map.current);
+        }
       });
     });
   }, [results, shouldColorZipCode]);
